requirements.txt
streamlit==1.29.0
PyPDF2==3.0.1
python-docx==1.1.0
transformers==4.36.2
torch==2.1.0